+++ internal/gcp/gcp_test.go
@@ -3,7 +3,6 @@ package gcp
 import (
 	"fmt"
 	"testing"
-	"time"
 
 	"github.com/hashicorp/terraform/states/statefile"
 	"github.com/kyma-incubator/hydroform/internal/operator/mocks"
@@ -92,16 +91,7 @@ func TestValidateInputs(t *testing.T) {
 }
 
 func TestLoadConfigurations(t *testing.T) {
-
-	timeouts := types.Timeouts{
-		Create: 10 * time.Minute,
-		Update: 20 * time.Minute,
-		Delete: 30 * time.Minute,
-	}
-
-	g := &gcpProvisioner{
-		timeouts: timeouts,
-	}
+	g := &gcpProvisioner{}
 
 	cluster := &types.Cluster{
 		CPU:               1,
@@ -134,9 +124,6 @@ func TestLoadConfigurations(t *testing.T) {
 	require.Equal(t, cluster.KubernetesVersion, config["kubernetes_version"])
 	require.Equal(t, cluster.Location, config["location"])
 	require.Equal(t, provider.ProjectName, config["project"])
-	require.Equal(t, timeouts.Create, config["create_timeout"])
-	require.Equal(t, timeouts.Update, config["update_timeout"])
-	require.Equal(t, timeouts.Delete, config["delete_timeout"])
 
 	for k, v := range provider.CustomConfigurations {
 		require.Equal(t, v, config[k], fmt.Sprintf("Custom config %s is incorrect", k))
