+++ internal/operator/terraform/timeouts_test.go
@@ -1,4 +1,4 @@
-package config
+package terraform
 
 import (
 	"testing"
