+++ kubeconfig.yaml
@@ -1,0 +1,22 @@
+---
+apiVersion: v1
+kind: Config
+current-context: shoot--sg-test--hydro-gcp2
+clusters:
+- name: shoot--sg-test--hydro-gcp2
+  cluster:
+    certificate-authority-data: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUMxekNDQWIrZ0F3SUJBZ0lRS2p1RzMvRjgxUzAwaW1RbEpFbDRaVEFOQmdrcWhraUc5dzBCQVFzRkFEQVYKTVJNd0VRWURWUVFERXdwcmRXSmxjbTVsZEdWek1CNFhEVEU1TVRJeE1URXpNekEwTlZvWERUSTVNVEl4TVRFegpNekEwTlZvd0ZURVRNQkVHQTFVRUF4TUthM1ZpWlhKdVpYUmxjekNDQVNJd0RRWUpLb1pJaHZjTkFRRUJCUUFECmdnRVBBRENDQVFvQ2dnRUJBTVFkb0FUZGovR0EySktGeWVWZllwTmtKQU9PSVY5YTdqeENnNVV1cXFUNWNrckEKTjFmcFh5T0ZSanNmOGkvNTZWUWp4WnlPZkEyQTZ2T2ZxVjMrYzVtd2duZUpLWjU0ODZHb1NXdUtMTm5qaGZmMgpqRkE0eTRzM2ZaUk1aY25vaDdEMmwwbGFNbjZxWkVGQzhhV3RzN3NadkNKbG1DSy8xMmp4bE9KbEtLcTZGSVRLCmt2cXR1NWd4RWVtZWl5dXk1MjN2dHhNVitIL2EvUU5wdlpZQy9MNGtVZFlkQ3pXWEZleXU2WE15MExLeERNK0kKTFB5a2RUUVdPaUQvRFJBbHR4SCtEZ1IxOXRXUDh6OHJJdk90Z3dMdlZlWEJadzIyZSt6MGRaSTBYT1c4TkduSworeW9YUzg1a0poUlg1VHBjeUl4cUJibnFpOFhTdFZpa3h5NDFWSDBDQXdFQUFhTWpNQ0V3RGdZRFZSMFBBUUgvCkJBUURBZ0dtTUE4R0ExVWRFd0VCL3dRRk1BTUJBZjh3RFFZSktvWklodmNOQVFFTEJRQURnZ0VCQUZ6clQvaVUKekZLekNEOTlGcjExS3VUN05wL29qVkx1ekFrYWlHaVF1NVZmNHUyOXZxSlhrc2lxWFhpQkdaUVZvWkJDMGljQworM1RSRUNlQkZPT3FETTRjNGVQY3BUWmRZa3NsWVh1dWd5VHoyMXI3ZnRuM2pVWk5RL0RBa21NMDYwdGo4UmJuClFJa2N3c2VyaTZsREp6b1JJellBR25NaUo2bXBpbklIdlNITFZaRVRNUHg3YjBkVnB5ZC9mamdMeUk0REJXMkkKOWsvWDkwaUdXaHgrSUtjZzRXbkpSMy9zWlNwRk90QVlkMTdSU2cwdTc4bFZaMHAybWFBT0E2aEZPV042Uy9uMApGZGJWZXdxUHNxbVVPYnUrVnBQc0RxbDFrcmRXbzlQd2pHZC9vak93V3VyYUhRaVhZVzUyYjBjcDRaSFpLdUN0Ci92K2pFRmJ3dVNVek5IND0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=
+    server: https://api.hydro-gcp2.sg-test.shoot.canary.k8s-hana.ondemand.com
+contexts:
+- name: shoot--sg-test--hydro-gcp2
+  context:
+    cluster: shoot--sg-test--hydro-gcp2
+    user: shoot--sg-test--hydro-gcp2-token
+users:
+- name: shoot--sg-test--hydro-gcp2-token
+  user:
+    token: V31StMf1YRvfTzdU0DHQZIODMtUZvqx25k5HIca8au15iNXfuD8vSzTcmj2suA7zQ7KB5KMouQEy0l6DQnHr8lV2llNl0FLvRbb0r2x9AqHGunXe3qvMjoE8c3y9lixq
+- name: shoot--sg-test--hydro-gcp2-basic-auth
+  user:
+    username: admin
+    password: DA8meV1L8e1qYUZg4xmFym2gvpsgu80f
